[screen] Update hardstatus string: Include more information; encapsulate inner hardstatus clearly
This was another experiment on 2016-11-04/-05. It's been deployed
to several hosts, too; its usefulness is potentially questionable,
due to information overflow, but perhaps it's better to give the
information so it can be looked at critically in, e.g., a screenshot,
later. (?)

If someone doesn't like it, he's supposed to use the commented-out
prior version instead.

Again, the documentation update has been added today, 2018-05-05.